refactor(ui): extract ownerFromJwt to shared lib/jwtUtils.ts
DRY fix - extracts duplicate JWT parsing logic to shared utility:

- Create lib/jwtUtils.ts with proper base64url decoding (RFC 4648)
- Update chat/send/route.ts to use shared ownerFromJwt
- Update chat/messages/route.ts to use shared ownerFromJwt
- Add component parameter for contextual error logging

The shared implementation fixes base64url encoding by replacing
'-' with '+' and '_' with '/' before decoding, and handles
padding correctly per RFC 4648.
